Making an executable works
In order to make everything function I had to move some of the files
around.

Moved Login/login_screen.py to the main.py
Moved main.py to Start/run.py

It is not possible to test the game from here, this is only meant for
making an executable file out of the game
