+++ .gitignore
@@ -4,3 +4,5 @@
 .idea/misc.xml
 *.htaccess
 *.pyc
+*Racer/build
+*Racer/dist
