+++ Racer/setup.py
@@ -3,7 +3,7 @@
 Game executable file
 """
 
-import cx_Freeze
+from cx_Freeze import setup, Executable
 import sys
 
 base = None
@@ -11,11 +11,15 @@ base = None
 if sys.platform == 'win32':
     base = "Win32GUI"
 
-executables = [cx_Freeze.Executable("distribution.py", base=base)]
+executables = [Executable("main.py", base=base)]
+packages = ["tkinter", "pygame", "pymysql", "bcrypt"]
+files = ["src/blue_car.png", "src/cars.png", "src/mainroad.png", "Music/racer.mp3",
+         "Background", "Buttons", "Car", "Login", "Menu", "Music", "Score", "Settings",
+         "Start", "Text", "README"]
 
-cx_Freeze.setup(
+setup(
     name="Racer",
-    options={"build_exe": {"packages":["tkinter", "pygame", "pymysql", "bcrypt"], "include_files": ["src/blue_car.png", "src/cars.png", "src/mainroad.png"]}},
+    options={"build_exe": {"packages": packages, "include_files": files}},
     version="0.01",
     description="Racer game",
     executables=executables
